+++ OggConverter/Class/Settings.cs
@@ -5,13 +5,19 @@ namespace OggConverter
     class Settings
     {
         public static bool RemoveMP3 { get; set; }
+        public static bool CloseAfterConversion { get; set; }
+        public static bool LaunchAfterConversion { get; set; }
+        public static bool NoSteam { get; set; }
 
         public Settings()
         {
-            RemoveMP3 = Load("RemoveMP3");
+            RemoveMP3 = Bool("RemoveMP3");
+            CloseAfterConversion = Bool("CloseAfterConversion");
+            LaunchAfterConversion = Bool("LaunchAfterConversion");
+            NoSteam = Bool("NoSteam");
         }
 
-        bool Load(string KeyName)
+        bool Bool(string KeyName)
         {
             using (RegistryKey Key = Registry.CurrentUser.CreateSubKey(@"SOFTWARE\MSCOGG", true))
             {
@@ -35,12 +41,27 @@ namespace OggConverter
                 }
             }
             return false;
-        }    
+        } 
+
+        string Variable(string KeyName)
+        {
+            using (RegistryKey Key = Registry.CurrentUser.CreateSubKey(@"SOFTWARE\MSCOGG", true))
+            {
+                object Value = Key.GetValue(KeyName);
+
+                if (Value != null)
+                {
+                    return Value.ToString();
+                }
+                else
+                    return "None";
+            }
+        }
     }
 
     public class ChangeSettings
     {
-        public static void ChangeBool(string KeyName)
+        public static void Bool(string KeyName)
         {
             using (RegistryKey Key = Registry.CurrentUser.CreateSubKey(@"SOFTWARE\MSCOGG", true))
             {
@@ -63,6 +84,28 @@ namespace OggConverter
                     Key.SetValue(KeyName, "true");
                 }
             }
+            new Settings();
+        }
+
+        // Always sets Bool to false
+        public static void BoolFalse(string KeyName)
+        {
+            using (RegistryKey Key = Registry.CurrentUser.CreateSubKey(@"SOFTWARE\MSCOGG", true))
+            {
+                object Value = Key.GetValue(KeyName);
+                Key.SetValue(KeyName, "false");
+            }
+            new Settings();
+        }
+
+        public static void Variable(string KeyName, string val)
+        {
+            using (RegistryKey Key = Registry.CurrentUser.CreateSubKey(@"SOFTWARE\MSCOGG", true))
+            {
+                object Value = Key.GetValue(KeyName);
+                Key.SetValue(KeyName, val);
+            }
+            new Settings();
         }
     }
 }